feat(operator): adds an operator that can create and bootstrap
This change represent a move from manually applying yaml resource
definitions to use the operator to automatically apply resources.

Generally speaking each yaml file has been converted into a rust file
that defines the same resources however now with the ability to
programmatically customize parts of each resource.

The simulation logic has been temporarily disabled because the operator
made a paradigm shift in how the runner is invoked. Previously the
runner assumed a deterministic set of nodes and DNS addresses to those
nodes. However with the operator things are a bit more dynamic and its
also easier to _assemble_ the set of nodes as opposed to assume their
existence deterministically. The bootstrap runner command has been
updated according to this new pattern. However to keep the PR smaller
the simulation command has not been updated.