fix: add tests for operator
These tests follow a basic pattern. First the controller is basically
responsible for making API calls to the k8s API. So to test it we
use a combination of mocks, stubs and expect_test to build a basic
test framework. The mock/stubs capture the requests made to the API
and return responses. The `expect_test` crate is used to assert the
correctness of the requests while making it easy to update the tests
when the code itself is changed. Additionally an `ExpectPatch` concept
is added that allows for  patching the default stubs instead of
reproducing their entire contents for each test.