feat: add network TTL
Now that we are programmatically creating networks (see ceramic-tests)
we want to make sure we are not leaving lots of unused resources around.

With this feature the test driver can create networks with a generous
TTL of a few hours (when we expect tests to run in a few minutes).
If for some reason the test driver fails to clean up the network itself,
eventually the operator will clean it up.

This feature remains optional so manually created networks are not
affected.