feat: devMode setting to override resource requests/limits
I implemented for the network using a global variable, which is ugly, but it applies to the entire network and doesn't seem totally wrong.
I think it could be done by passing things down, but it will probably require some refactoring to simplify CAS and other stateful spec declaration.
I considered adding this in the default impls, but I don't like having a parameterized default, even though it might make it more clear than having to call `Self::network_default()`