feat: move monitoring resources to network
There is now a `monitoring` flag on networks when true deploys
opentelemetry, jaeger and prometheus. Opentelemetry can been changed to
no longer scrape targets, rather it only listens for pushed metrics.

The simulation no longer deploys any of these resources. However the
simulation does check that they exist and are running in order to start
a simulation. This means that on our network were we intend to run
simulations we need to set `monitoring=true` on the network.

Additionally logic was added to the simulate logic inside the runner to
report the minimum rps of any peer after a simulation completes.

With these two changes we should be able to capture the min_peer_rps as
the most important metrics from simulation runs and persist that metric
in our cloud systems.