feat: expose two separate metrics endpoints
With this change not only have the monitoring services been moved to be
part of the network but now simulation metrics can be collected
separately from the rest of the network metrics.

The prometheus endpoint will scrape all metrics, while external systems
can scrape otel:9465 to get only the simulation metrics.