feat: expose prometheus metrics from operator
Now the operator exposes a prometheus scrape endpoint. It does not try
to push its metrics over OTLP. This operator is long lived so this
should not be a problem.

The runner continues to push its metrics over OTLP which is important
since it is short lived.

However I was not able to get both prometheus scrape and OTLP push
working on the same process. So the operator only supports prometheus
and the runner only support OTLP push.