feat: configure ceramic to use postgres
This changes the ceramic pod configuration to include a postgres pod
(copied from the CAS postgres pod) and configures Ceramic to use the
postgres pod.

This update is breaking because the `ceramic` stateful_set is changed in
ways that k8s doesn't allow without deleting and recreating the
stateful_set. Therefore we will need to determine a rollout strategy.
Can we simply rebuild all networks? If not we need to investigate how to
update.