fix: no default otel endpoint for the operator (#154)
* fix: no default otel endpoint for the operator

The operator no longer has a default opentelemetry endpoint and will
only collect and send traces if an endpoint is configured.

This might address the memory leak but either way its a nice quality of
life improvement as we have not been collecting traces from the operator
anyway.

* fix: enable logs even if otlp tracing export is disabled

With this change logs will still be printed to STDOUT even if no OTLP
endpoint is configured.